pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-media-langgraph"
version = "0.1.0"
description = "A social media agent using LangGraph and FastHTML to generate and schedule posts"
requires-python = ">=3.11"
authors = [
    {name = "Social Media Agent Team", email = "social@predictivelabs.co.uk"}
]
license = {text = "MIT"}

dependencies = [
    "python-fasthtml>=0.12.0",
    "langgraph>=0.1.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.1.0",
    "langchain-core>=0.1.0",
    "python-dotenv>=1.0.0",
    "firecrawl-py>=1.0.0",
    "arcade-sdk>=1.0.0",
    "httpx>=0.24.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "uvicorn>=0.24.0",
    "starlette>=0.35.0",
    "python-multipart>=0.0.6",
    "aiofiles>=23.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
